enable Primer3 as an alternative primer design tool
This commit adds generation of Primer3 (v2+) command-lines, and
output parsing. The pdp primer3 subcommand behaves like the
pdp eprimer3 subcommand. Both generate JSON output that should be
usable in downstream steps, almost interchangeably.
