+++ diagnostic_primers/eprimer3.py
@@ -44,26 +44,10 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 THE SOFTWARE.
 """
 
-import json
 import os
 
 from Bio import SeqIO
 from Bio.Emboss.Applications import Primer3Commandline
-from Bio.Emboss import Primer3
-from Bio.Seq import Seq
-from Bio.SeqRecord import SeqRecord
-
-
-class PrimersEncoder(json.JSONEncoder):
-    """JSON encoder for Primer3.Primers objects."""
-
-    def default(self, obj):
-        if not isinstance(obj, Primer3.Primers):
-            return json.JSONEncoder.default(self, obj)
-
-        # Convert complex Primer3.Primers object to serialisable dictionary
-        # and return
-        return obj.__dict__
 
 
 def build_commands(collection, eprimer3_exe, eprimer3_dir, argdict=None):
@@ -114,203 +98,3 @@ def build_command(eprimer3_exe, seqfile, filestem, argdict=None):
                 setattr(cline, arg, val)
     setattr(cline, "prange", "%d-%d" % tuple(prange))
     return cline
-
-
-def load_primers(infname, fmt="eprimer3", noname=False):
-    """Load primers from a file.
-
-    The function can load JSON or ePrimer3 files - ePrimer3 by default.
-    """
-    if fmt in ("ep3", "eprimer3"):
-        return __load_primers_eprimer3(infname, noname)
-    elif fmt in ("json",):
-        return __load_primers_json(infname)
-
-
-def __load_primers_eprimer3(infname, noname=False):
-    """Loads and names primers from the passed ePrimer3 file. Returns JSON.
-
-    noname  - Boolean flag. Unless set to true, each primer receives
-              a unique name based on the input filename.
-    """
-    # Load primers with Biopython. This does not respect a 'name' attribute,
-    # as the bare ePrimer3 files don't provide names for primer sets.
-    with open(infname, "r") as primerfh:
-        ep3primers = Primer3.read(primerfh).primers
-
-    # Create unique identifier for each primer set, based on the input
-    # filename
-    if not noname:
-        for idx, primer in enumerate(ep3primers, 1):
-            stem = os.path.splitext(os.path.split(infname)[-1])[0]
-            primer.name = "%s_primer_%05d" % (stem, idx)
-
-    # Return primers
-    return ep3primers
-
-
-def __load_primers_json(infname):
-    """Loads and returns primers from a JSON file."""
-    primers = []
-    with open(infname, "r") as primerfh:
-        for pdata in json.load(primerfh):
-            primer = Primer3.Primers()
-            for k, v in pdata.items():
-                setattr(primer, k, v)
-            primers.append(primer)
-    return primers
-
-
-def write_primers(primers, outfilename, fmt="fasta"):
-    """Write Primer3.Primers to file.
-
-    primers      - collection of Biopython primer objects
-    outfilename  - path to output file
-    format       - sequence format to write
-
-    TODO: distribution dictionary
-    """
-    # Ensure output directory exists
-    outdir = os.path.join(*os.path.split(outfilename)[:-1])
-    os.makedirs(outdir, exist_ok=True)
-
-    # Order primers before writing
-    primers = [_[1] for _ in sorted([(primer.name, primer) for primer in primers])]
-    if fmt in ("json",):
-        __write_primers_json(primers, outfilename)
-    elif fmt in ("ep3", "eprimer3"):
-        __write_primers_eprimer3(primers, outfilename)
-    elif fmt in ("tsv", "tab"):
-        __write_primers_tsv(primers, outfilename)
-    elif fmt in ("bed"):
-        __write_primers_bed(primers, outfilename)
-    else:
-        __write_primers_seqio(primers, outfilename, fmt)
-
-
-def __write_primers_seqio(primers, outfilename, fmt):
-    """Write primers to file, using SeqIO.
-
-    Returns the number of records written
-    """
-    seqrecords = []
-
-    for primer in primers:
-        seqrecords.append(
-            SeqRecord(Seq(primer.forward_seq), id=primer.name + "_fwd", description="")
-        )
-        seqrecords.append(
-            SeqRecord(Seq(primer.reverse_seq), id=primer.name + "_rev", description="")
-        )
-        if len(primer.internal_seq):  # This is '' id no oligo
-            seqrecords.append(
-                SeqRecord(
-                    Seq(primer.internal_seq), id=primer.name + "_int", description=""
-                )
-            )
-
-    return SeqIO.write(seqrecords, outfilename, fmt)
-
-
-def __write_primers_tsv(primers, outfname):
-    """Write primers to file in three-column tab-separated format.
-
-    This is required for primersearch input with EMBOSS
-    """
-    # Don't use more than one newline at the end of the header, or
-    # else primersearch treats the blank line as a primer!
-    header = (
-        "\n".join(["# EPRIMER3 PRIMERS %s" % outfname, "# Name       FWD        REV"])
-        + "\n"
-    )
-    with open(outfname, "w") as outfh:
-        outfh.write(header)
-        for primer in primers:
-            outfh.write(
-                "\t".join([primer.name, primer.forward_seq, primer.reverse_seq]) + "\n"
-            )
-
-
-def __write_primers_eprimer3(primers, outfname):
-    """Write Primer3 primer objects in ePrimer3 format (Extended)."""
-    header = (
-        "\n".join(
-            [
-                "# EPRIMER3 PRIMERS %s " % outfname,
-                "#                      Start  Len   Tm     " + "GC%   Sequence",
-            ]
-        )
-        + "\n"
-    )
-
-    with open(outfname, "w") as outfh:
-        outfh.write(header)
-        for idx, primer in enumerate(primers, 1):
-            outfh.write("# %s\n" % primer.name)
-            outfh.write("%-4d PRODUCT SIZE: %d\n" % (idx, primer.size))
-            outfh.write(
-                "     FORWARD PRIMER  %-9d  %-3d  %.02f  %.02f  %s\n"
-                % (
-                    primer.forward_start,
-                    primer.forward_length,
-                    primer.forward_tm,
-                    primer.forward_gc,
-                    primer.forward_seq,
-                )
-            )
-            outfh.write(
-                "     REVERSE PRIMER  %-9d  %-3d  %.02f  %.02f  %s\n"
-                % (
-                    primer.reverse_start,
-                    primer.reverse_length,
-                    primer.reverse_tm,
-                    primer.reverse_gc,
-                    primer.reverse_seq,
-                )
-            )
-            if hasattr(primer, "internal_start"):
-                outfh.write(
-                    "     INTERNAL OLIGO  "
-                    + "%-9d  %-3d  %.02f  %.02f  %s\n"
-                    % (
-                        primer.internal_start,
-                        primer.internal_length,
-                        primer.internal_tm,
-                        primer.internal_gc,
-                        primer.internal_seq,
-                    )
-                )
-            outfh.write("\n" * 3)
-
-
-def __write_primers_json(primers, outfname):
-    """Write Primer3 primer objects in JSON format."""
-    with open(outfname, "w") as ofh:
-        json.dump(primers, ofh, cls=PrimersEncoder)
-
-
-def __write_primers_bed(primers, outfname):
-    """Write Primer3 primer objects in BED format."""
-    with open(outfname, "w") as outfh:
-        sourceids = {}
-        for primer in primers:
-            try:
-                source_id = sourceids[primer.source]  #  lazily acquire primer source ID
-            except KeyError:
-                source_id = sourceids.setdefault(
-                    primer.source, load_fasta_id(primer.source)
-                )
-            outfh.write(
-                "{}\t{}\t{}\t{}\n".format(
-                    source_id,
-                    primer.forward_start,
-                    primer.reverse_start - 1,
-                    primer.name,
-                )
-            )
-
-
-def load_fasta_id(fname):
-    """Return the identifier from the passed FASTA file."""
-    with open(fname, "r") as ifh:
-        return SeqIO.read(ifh, "fasta").id
