+++ tests/test_targets/pdp_primer3/subset/GCF_000011605.1_named.bed
@@ -1,0 +1,10 @@
+gi|49609491|emb|BX950851.1|	4144856	4144934	GCF_000011605.1_primer_00001
+gi|49609491|emb|BX950851.1|	4567898	4567976	GCF_000011605.1_primer_00002
+gi|49609491|emb|BX950851.1|	2094087	2094165	GCF_000011605.1_primer_00003
+gi|49609491|emb|BX950851.1|	150787	150865	GCF_000011605.1_primer_00004
+gi|49609491|emb|BX950851.1|	1552064	1552142	GCF_000011605.1_primer_00005
+gi|49609491|emb|BX950851.1|	1125835	1125913	GCF_000011605.1_primer_00006
+gi|49609491|emb|BX950851.1|	2390734	2390812	GCF_000011605.1_primer_00007
+gi|49609491|emb|BX950851.1|	3539230	3539308	GCF_000011605.1_primer_00008
+gi|49609491|emb|BX950851.1|	882019	882097	GCF_000011605.1_primer_00009
+gi|49609491|emb|BX950851.1|	1573548	1573626	GCF_000011605.1_primer_00010
