+++ tests/test_targets/pdp_primer3/subset/GCF_000011605.1_named.eprimer3
@@ -1,0 +1,82 @@
+# EPRIMER3 PRIMERS tests/test_output/pdp_primer3/subset/GCF_000011605.1_named.eprimer3 
+#                      Start  Len   Tm     GC%   Sequence
+# GCF_000011605.1_primer_00001
+1    PRODUCT SIZE: 100
+     FORWARD PRIMER  4144856    20   59.00  55.00  CCTCTCCGATTTGTTCAGCG
+     REVERSE PRIMER  4144935    20   59.00  50.00  AAAGGGCCTGAGCTATTGGT
+     INTERNAL OLIGO  0          0    0.00  0.00  
+
+
+
+# GCF_000011605.1_primer_00002
+2    PRODUCT SIZE: 100
+     FORWARD PRIMER  4567898    20   59.00  50.00  ACATCACGCAGTCGCAATAC
+     REVERSE PRIMER  4567977    20   58.99  50.00  ATTTCTGCGACATACGGTGC
+     INTERNAL OLIGO  0          0    0.00  0.00  
+
+
+
+# GCF_000011605.1_primer_00003
+3    PRODUCT SIZE: 100
+     FORWARD PRIMER  2094087    20   59.00  55.00  CGCGCTAGGGTCAAAGTTAC
+     REVERSE PRIMER  2094166    20   58.99  50.00  TATTGTTGATCTGGCACGCG
+     INTERNAL OLIGO  0          0    0.00  0.00  
+
+
+
+# GCF_000011605.1_primer_00004
+4    PRODUCT SIZE: 100
+     FORWARD PRIMER  150787     20   58.99  50.00  CGATTCGCGGTTGAATTCCT
+     REVERSE PRIMER  150866     20   59.01  55.00  CCTCGATGCTCACGTCAAAG
+     INTERNAL OLIGO  0          0    0.00  0.00  
+
+
+
+# GCF_000011605.1_primer_00005
+5    PRODUCT SIZE: 100
+     FORWARD PRIMER  1552064    20   59.01  55.00  GGTCGTCGTGAGCTGAAATC
+     REVERSE PRIMER  1552143    20   58.99  50.00  GTGTGATAATCGGCGCTGTT
+     INTERNAL OLIGO  0          0    0.00  0.00  
+
+
+
+# GCF_000011605.1_primer_00006
+6    PRODUCT SIZE: 100
+     FORWARD PRIMER  1125835    20   59.00  50.00  TGCTGACGAATCGCAATCAG
+     REVERSE PRIMER  1125914    20   58.99  50.00  ACGGTGAACACAAAGTGAGC
+     INTERNAL OLIGO  0          0    0.00  0.00  
+
+
+
+# GCF_000011605.1_primer_00007
+7    PRODUCT SIZE: 100
+     FORWARD PRIMER  2390734    20   59.02  50.00  AAGGAGAAGACCGATGGCAA
+     REVERSE PRIMER  2390813    20   59.00  50.00  ATCCAGCCTGAAATCCACCA
+     INTERNAL OLIGO  0          0    0.00  0.00  
+
+
+
+# GCF_000011605.1_primer_00008
+8    PRODUCT SIZE: 100
+     FORWARD PRIMER  3539230    20   58.98  55.00  GGTCATGATGAAACTCGGCC
+     REVERSE PRIMER  3539309    20   59.00  50.00  TGCCAGCCCAAGGAAGATAA
+     INTERNAL OLIGO  0          0    0.00  0.00  
+
+
+
+# GCF_000011605.1_primer_00009
+9    PRODUCT SIZE: 100
+     FORWARD PRIMER  882019     20   58.98  55.00  GCGTGATTAGCCAAGAAGGG
+     REVERSE PRIMER  882098     20   59.00  50.00  GTCGCATCTGCACGAGAAAT
+     INTERNAL OLIGO  0          0    0.00  0.00  
+
+
+
+# GCF_000011605.1_primer_00010
+10   PRODUCT SIZE: 100
+     FORWARD PRIMER  1573548    20   58.99  55.00  GCAGGATTAGCCACTGTTCG
+     REVERSE PRIMER  1573627    20   59.01  55.00  GTACGCAACGTCATCCTGTC
+     INTERNAL OLIGO  0          0    0.00  0.00  
+
+
+
