+++ tests/test_targets/pdp_primer3/subset/GCF_000011605.1_named.json
@@ -1,0 +1,1 @@
+[{"size": 100, "forward_seq": "CCTCTCCGATTTGTTCAGCG", "forward_start": 4144856, "forward_length": 20, "forward_tm": 58.998, "forward_gc": 55.0, "reverse_seq": "AAAGGGCCTGAGCTATTGGT", "reverse_start": 4144935, "reverse_length": 20, "reverse_tm": 58.997, "reverse_gc": 50.0, "internal_seq": "", "internal_start": 0, "internal_length": 0, "internal_tm": 0.0, "internal_gc": 0.0, "name": "GCF_000011605.1_primer_00001", "source": "tests/test_input/sequences/GCF_000011605.1.fasta", "sourcename": "Pba_SCRI1043"}, {"size": 100, "forward_seq": "ACATCACGCAGTCGCAATAC", "forward_start": 4567898, "forward_length": 20, "forward_tm": 59.001, "forward_gc": 50.0, "reverse_seq": "ATTTCTGCGACATACGGTGC", "reverse_start": 4567977, "reverse_length": 20, "reverse_tm": 58.993, "reverse_gc": 50.0, "internal_seq": "", "internal_start": 0, "internal_length": 0, "internal_tm": 0.0, "internal_gc": 0.0, "name": "GCF_000011605.1_primer_00002", "source": "tests/test_input/sequences/GCF_000011605.1.fasta", "sourcename": "Pba_SCRI1043"}, {"size": 100, "forward_seq": "CGCGCTAGGGTCAAAGTTAC", "forward_start": 2094087, "forward_length": 20, "forward_tm": 59.003, "forward_gc": 55.0, "reverse_seq": "TATTGTTGATCTGGCACGCG", "reverse_start": 2094166, "reverse_length": 20, "reverse_tm": 58.994, "reverse_gc": 50.0, "internal_seq": "", "internal_start": 0, "internal_length": 0, "internal_tm": 0.0, "internal_gc": 0.0, "name": "GCF_000011605.1_primer_00003", "source": "tests/test_input/sequences/GCF_000011605.1.fasta", "sourcename": "Pba_SCRI1043"}, {"size": 100, "forward_seq": "CGATTCGCGGTTGAATTCCT", "forward_start": 150787, "forward_length": 20, "forward_tm": 58.993, "forward_gc": 50.0, "reverse_seq": "CCTCGATGCTCACGTCAAAG", "reverse_start": 150866, "reverse_length": 20, "reverse_tm": 59.005, "reverse_gc": 55.0, "internal_seq": "", "internal_start": 0, "internal_length": 0, "internal_tm": 0.0, "internal_gc": 0.0, "name": "GCF_000011605.1_primer_00004", "source": "tests/test_input/sequences/GCF_000011605.1.fasta", "sourcename": "Pba_SCRI1043"}, {"size": 100, "forward_seq": "GGTCGTCGTGAGCTGAAATC", "forward_start": 1552064, "forward_length": 20, "forward_tm": 59.007, "forward_gc": 55.0, "reverse_seq": "GTGTGATAATCGGCGCTGTT", "reverse_start": 1552143, "reverse_length": 20, "reverse_tm": 58.993, "reverse_gc": 50.0, "internal_seq": "", "internal_start": 0, "internal_length": 0, "internal_tm": 0.0, "internal_gc": 0.0, "name": "GCF_000011605.1_primer_00005", "source": "tests/test_input/sequences/GCF_000011605.1.fasta", "sourcename": "Pba_SCRI1043"}, {"size": 100, "forward_seq": "TGCTGACGAATCGCAATCAG", "forward_start": 1125835, "forward_length": 20, "forward_tm": 58.998, "forward_gc": 50.0, "reverse_seq": "ACGGTGAACACAAAGTGAGC", "reverse_start": 1125914, "reverse_length": 20, "reverse_tm": 58.986, "reverse_gc": 50.0, "internal_seq": "", "internal_start": 0, "internal_length": 0, "internal_tm": 0.0, "internal_gc": 0.0, "name": "GCF_000011605.1_primer_00006", "source": "tests/test_input/sequences/GCF_000011605.1.fasta", "sourcename": "Pba_SCRI1043"}, {"size": 100, "forward_seq": "AAGGAGAAGACCGATGGCAA", "forward_start": 2390734, "forward_length": 20, "forward_tm": 59.016, "forward_gc": 50.0, "reverse_seq": "ATCCAGCCTGAAATCCACCA", "reverse_start": 2390813, "reverse_length": 20, "reverse_tm": 58.998, "reverse_gc": 50.0, "internal_seq": "", "internal_start": 0, "internal_length": 0, "internal_tm": 0.0, "internal_gc": 0.0, "name": "GCF_000011605.1_primer_00007", "source": "tests/test_input/sequences/GCF_000011605.1.fasta", "sourcename": "Pba_SCRI1043"}, {"size": 100, "forward_seq": "GGTCATGATGAAACTCGGCC", "forward_start": 3539230, "forward_length": 20, "forward_tm": 58.983, "forward_gc": 55.0, "reverse_seq": "TGCCAGCCCAAGGAAGATAA", "reverse_start": 3539309, "reverse_length": 20, "reverse_tm": 59.001, "reverse_gc": 50.0, "internal_seq": "", "internal_start": 0, "internal_length": 0, "internal_tm": 0.0, "internal_gc": 0.0, "name": "GCF_000011605.1_primer_00008", "source": "tests/test_input/sequences/GCF_000011605.1.fasta", "sourcename": "Pba_SCRI1043"}, {"size": 100, "forward_seq": "GCGTGATTAGCCAAGAAGGG", "forward_start": 882019, "forward_length": 20, "forward_tm": 58.983, "forward_gc": 55.0, "reverse_seq": "GTCGCATCTGCACGAGAAAT", "reverse_start": 882098, "reverse_length": 20, "reverse_tm": 58.998, "reverse_gc": 50.0, "internal_seq": "", "internal_start": 0, "internal_length": 0, "internal_tm": 0.0, "internal_gc": 0.0, "name": "GCF_000011605.1_primer_00009", "source": "tests/test_input/sequences/GCF_000011605.1.fasta", "sourcename": "Pba_SCRI1043"}, {"size": 100, "forward_seq": "GCAGGATTAGCCACTGTTCG", "forward_start": 1573548, "forward_length": 20, "forward_tm": 58.99, "forward_gc": 55.0, "reverse_seq": "GTACGCAACGTCATCCTGTC", "reverse_start": 1573627, "reverse_length": 20, "reverse_tm": 59.01, "reverse_gc": 55.0, "internal_seq": "", "internal_start": 0, "internal_length": 0, "internal_tm": 0.0, "internal_gc": 0.0, "name": "GCF_000011605.1_primer_00010", "source": "tests/test_input/sequences/GCF_000011605.1.fasta", "sourcename": "Pba_SCRI1043"}]
