+++ tests/test_targets/pdp_primer3/subset/GCF_000740965.1_concat_named.bed
@@ -1,0 +1,10 @@
+Pba_21A_concatenated	4334254	4334332	GCF_000740965.1_concat_primer_00001
+Pba_21A_concatenated	1235051	1235129	GCF_000740965.1_concat_primer_00002
+Pba_21A_concatenated	3015101	3015179	GCF_000740965.1_concat_primer_00003
+Pba_21A_concatenated	154018	154096	GCF_000740965.1_concat_primer_00004
+Pba_21A_concatenated	1533186	1533264	GCF_000740965.1_concat_primer_00005
+Pba_21A_concatenated	475836	475914	GCF_000740965.1_concat_primer_00006
+Pba_21A_concatenated	2724077	2724155	GCF_000740965.1_concat_primer_00007
+Pba_21A_concatenated	3462840	3462918	GCF_000740965.1_concat_primer_00008
+Pba_21A_concatenated	763849	763927	GCF_000740965.1_concat_primer_00009
+Pba_21A_concatenated	4069992	4070070	GCF_000740965.1_concat_primer_00010
