+++ tests/test_targets/pdp_primer3/subset/GCF_000740965.1_concat_named.eprimer3
@@ -1,0 +1,82 @@
+# EPRIMER3 PRIMERS tests/test_output/pdp_primer3/subset/GCF_000740965.1_concat_named.eprimer3 
+#                      Start  Len   Tm     GC%   Sequence
+# GCF_000740965.1_concat_primer_00001
+1    PRODUCT SIZE: 100
+     FORWARD PRIMER  4334254    20   59.00  55.00  CCTCTCCGATTTGTTCAGCG
+     REVERSE PRIMER  4334333    20   59.00  50.00  AAAGGGCCTGAGCTATTGGT
+     INTERNAL OLIGO  0          0    0.00  0.00  
+
+
+
+# GCF_000740965.1_concat_primer_00002
+2    PRODUCT SIZE: 100
+     FORWARD PRIMER  1235051    20   59.00  50.00  GCTTCATCTTGTAGCGCGAA
+     REVERSE PRIMER  1235130    20   59.01  55.00  CTGTTGTACCGCTGACGATG
+     INTERNAL OLIGO  0          0    0.00  0.00  
+
+
+
+# GCF_000740965.1_concat_primer_00003
+3    PRODUCT SIZE: 100
+     FORWARD PRIMER  3015101    20   58.99  50.00  TATTGTTGATCTGGCACGCG
+     REVERSE PRIMER  3015180    20   59.00  55.00  CGCGCTAGGGTCAAAGTTAC
+     INTERNAL OLIGO  0          0    0.00  0.00  
+
+
+
+# GCF_000740965.1_concat_primer_00004
+4    PRODUCT SIZE: 100
+     FORWARD PRIMER  154018     20   58.99  50.00  CGATTCGCGGTTGAATTCCT
+     REVERSE PRIMER  154097     20   59.01  55.00  CCTCGATGCTCACGTCAAAG
+     INTERNAL OLIGO  0          0    0.00  0.00  
+
+
+
+# GCF_000740965.1_concat_primer_00005
+5    PRODUCT SIZE: 100
+     FORWARD PRIMER  1533186    20   59.01  55.00  GGTCGTCGTGAGCTGAAATC
+     REVERSE PRIMER  1533265    20   58.99  50.00  GTGTGATAATCGGCGCTGTT
+     INTERNAL OLIGO  0          0    0.00  0.00  
+
+
+
+# GCF_000740965.1_concat_primer_00006
+6    PRODUCT SIZE: 100
+     FORWARD PRIMER  475836     20   59.00  55.00  CAGTCGGGTAAGCATCGTTG
+     REVERSE PRIMER  475915     20   59.02  55.00  GGTCTGGAACTGATGGGTCA
+     INTERNAL OLIGO  0          0    0.00  0.00  
+
+
+
+# GCF_000740965.1_concat_primer_00007
+7    PRODUCT SIZE: 100
+     FORWARD PRIMER  2724077    20   59.00  50.00  ATCCAGCCTGAAATCCACCA
+     REVERSE PRIMER  2724156    20   59.02  50.00  AAGGAGAAGACCGATGGCAA
+     INTERNAL OLIGO  0          0    0.00  0.00  
+
+
+
+# GCF_000740965.1_concat_primer_00008
+8    PRODUCT SIZE: 100
+     FORWARD PRIMER  3462840    20   58.98  55.00  GGTCATGATGAAACTCGGCC
+     REVERSE PRIMER  3462919    20   59.00  50.00  TGCCAGCCCAAGGAAGATAA
+     INTERNAL OLIGO  0          0    0.00  0.00  
+
+
+
+# GCF_000740965.1_concat_primer_00009
+9    PRODUCT SIZE: 100
+     FORWARD PRIMER  763849     20   58.98  55.00  GCGTGATTAGCCAAGAAGGG
+     REVERSE PRIMER  763928     20   59.00  50.00  GTCGCATCTGCACGAGAAAT
+     INTERNAL OLIGO  0          0    0.00  0.00  
+
+
+
+# GCF_000740965.1_concat_primer_00010
+10   PRODUCT SIZE: 100
+     FORWARD PRIMER  4069992    20   59.01  50.00  GTGAAGCAAATCGCCAATGC
+     REVERSE PRIMER  4070071    20   58.99  55.00  CGGTTGTATAGGTTGCTGCC
+     INTERNAL OLIGO  0          0    0.00  0.00  
+
+
+
