+++ tests/test_targets/pdp_primer3/subset/GCF_000740965.1_concat_named.json
@@ -1,0 +1,1 @@
+[{"size": 100, "forward_seq": "CCTCTCCGATTTGTTCAGCG", "forward_start": 4334254, "forward_length": 20, "forward_tm": 58.998, "forward_gc": 55.0, "reverse_seq": "AAAGGGCCTGAGCTATTGGT", "reverse_start": 4334333, "reverse_length": 20, "reverse_tm": 58.997, "reverse_gc": 50.0, "internal_seq": "", "internal_start": 0, "internal_length": 0, "internal_tm": 0.0, "internal_gc": 0.0, "name": "GCF_000740965.1_concat_primer_00001", "source": "tests/test_output/pdp_config/GCF_000740965.1_concat.fas", "sourcename": "Pba_21A"}, {"size": 100, "forward_seq": "GCTTCATCTTGTAGCGCGAA", "forward_start": 1235051, "forward_length": 20, "forward_tm": 59.001, "forward_gc": 50.0, "reverse_seq": "CTGTTGTACCGCTGACGATG", "reverse_start": 1235130, "reverse_length": 20, "reverse_tm": 59.008, "reverse_gc": 55.0, "internal_seq": "", "internal_start": 0, "internal_length": 0, "internal_tm": 0.0, "internal_gc": 0.0, "name": "GCF_000740965.1_concat_primer_00002", "source": "tests/test_output/pdp_config/GCF_000740965.1_concat.fas", "sourcename": "Pba_21A"}, {"size": 100, "forward_seq": "TATTGTTGATCTGGCACGCG", "forward_start": 3015101, "forward_length": 20, "forward_tm": 58.994, "forward_gc": 50.0, "reverse_seq": "CGCGCTAGGGTCAAAGTTAC", "reverse_start": 3015180, "reverse_length": 20, "reverse_tm": 59.003, "reverse_gc": 55.0, "internal_seq": "", "internal_start": 0, "internal_length": 0, "internal_tm": 0.0, "internal_gc": 0.0, "name": "GCF_000740965.1_concat_primer_00003", "source": "tests/test_output/pdp_config/GCF_000740965.1_concat.fas", "sourcename": "Pba_21A"}, {"size": 100, "forward_seq": "CGATTCGCGGTTGAATTCCT", "forward_start": 154018, "forward_length": 20, "forward_tm": 58.993, "forward_gc": 50.0, "reverse_seq": "CCTCGATGCTCACGTCAAAG", "reverse_start": 154097, "reverse_length": 20, "reverse_tm": 59.005, "reverse_gc": 55.0, "internal_seq": "", "internal_start": 0, "internal_length": 0, "internal_tm": 0.0, "internal_gc": 0.0, "name": "GCF_000740965.1_concat_primer_00004", "source": "tests/test_output/pdp_config/GCF_000740965.1_concat.fas", "sourcename": "Pba_21A"}, {"size": 100, "forward_seq": "GGTCGTCGTGAGCTGAAATC", "forward_start": 1533186, "forward_length": 20, "forward_tm": 59.007, "forward_gc": 55.0, "reverse_seq": "GTGTGATAATCGGCGCTGTT", "reverse_start": 1533265, "reverse_length": 20, "reverse_tm": 58.993, "reverse_gc": 50.0, "internal_seq": "", "internal_start": 0, "internal_length": 0, "internal_tm": 0.0, "internal_gc": 0.0, "name": "GCF_000740965.1_concat_primer_00005", "source": "tests/test_output/pdp_config/GCF_000740965.1_concat.fas", "sourcename": "Pba_21A"}, {"size": 100, "forward_seq": "CAGTCGGGTAAGCATCGTTG", "forward_start": 475836, "forward_length": 20, "forward_tm": 59.001, "forward_gc": 55.0, "reverse_seq": "GGTCTGGAACTGATGGGTCA", "reverse_start": 475915, "reverse_length": 20, "reverse_tm": 59.016, "reverse_gc": 55.0, "internal_seq": "", "internal_start": 0, "internal_length": 0, "internal_tm": 0.0, "internal_gc": 0.0, "name": "GCF_000740965.1_concat_primer_00006", "source": "tests/test_output/pdp_config/GCF_000740965.1_concat.fas", "sourcename": "Pba_21A"}, {"size": 100, "forward_seq": "ATCCAGCCTGAAATCCACCA", "forward_start": 2724077, "forward_length": 20, "forward_tm": 58.998, "forward_gc": 50.0, "reverse_seq": "AAGGAGAAGACCGATGGCAA", "reverse_start": 2724156, "reverse_length": 20, "reverse_tm": 59.016, "reverse_gc": 50.0, "internal_seq": "", "internal_start": 0, "internal_length": 0, "internal_tm": 0.0, "internal_gc": 0.0, "name": "GCF_000740965.1_concat_primer_00007", "source": "tests/test_output/pdp_config/GCF_000740965.1_concat.fas", "sourcename": "Pba_21A"}, {"size": 100, "forward_seq": "GGTCATGATGAAACTCGGCC", "forward_start": 3462840, "forward_length": 20, "forward_tm": 58.983, "forward_gc": 55.0, "reverse_seq": "TGCCAGCCCAAGGAAGATAA", "reverse_start": 3462919, "reverse_length": 20, "reverse_tm": 59.001, "reverse_gc": 50.0, "internal_seq": "", "internal_start": 0, "internal_length": 0, "internal_tm": 0.0, "internal_gc": 0.0, "name": "GCF_000740965.1_concat_primer_00008", "source": "tests/test_output/pdp_config/GCF_000740965.1_concat.fas", "sourcename": "Pba_21A"}, {"size": 100, "forward_seq": "GCGTGATTAGCCAAGAAGGG", "forward_start": 763849, "forward_length": 20, "forward_tm": 58.983, "forward_gc": 55.0, "reverse_seq": "GTCGCATCTGCACGAGAAAT", "reverse_start": 763928, "reverse_length": 20, "reverse_tm": 58.998, "reverse_gc": 50.0, "internal_seq": "", "internal_start": 0, "internal_length": 0, "internal_tm": 0.0, "internal_gc": 0.0, "name": "GCF_000740965.1_concat_primer_00009", "source": "tests/test_output/pdp_config/GCF_000740965.1_concat.fas", "sourcename": "Pba_21A"}, {"size": 100, "forward_seq": "GTGAAGCAAATCGCCAATGC", "forward_start": 4069992, "forward_length": 20, "forward_tm": 59.006, "forward_gc": 50.0, "reverse_seq": "CGGTTGTATAGGTTGCTGCC", "reverse_start": 4070071, "reverse_length": 20, "reverse_tm": 58.986, "reverse_gc": 55.0, "internal_seq": "", "internal_start": 0, "internal_length": 0, "internal_tm": 0.0, "internal_gc": 0.0, "name": "GCF_000740965.1_concat_primer_00010", "source": "tests/test_output/pdp_config/GCF_000740965.1_concat.fas", "sourcename": "Pba_21A"}]
