+++ tests/test_targets/pdp_primer3/subset/GCF_001038685.1_concat_noambig_named.bed
@@ -1,0 +1,10 @@
+Pba_ICMP_1526_concatenated_noambig	914337	914415	GCF_001038685.1_concat_noambig_primer_00001
+Pba_ICMP_1526_concatenated_noambig	2533415	2533493	GCF_001038685.1_concat_noambig_primer_00002
+Pba_ICMP_1526_concatenated_noambig	3896478	3896556	GCF_001038685.1_concat_noambig_primer_00003
+Pba_ICMP_1526_concatenated_noambig	3467081	3467159	GCF_001038685.1_concat_noambig_primer_00004
+Pba_ICMP_1526_concatenated_noambig	2807481	2807559	GCF_001038685.1_concat_noambig_primer_00005
+Pba_ICMP_1526_concatenated_noambig	1210545	1210623	GCF_001038685.1_concat_noambig_primer_00006
+Pba_ICMP_1526_concatenated_noambig	445349	445427	GCF_001038685.1_concat_noambig_primer_00007
+Pba_ICMP_1526_concatenated_noambig	1585661	1585739	GCF_001038685.1_concat_noambig_primer_00008
+Pba_ICMP_1526_concatenated_noambig	3104191	3104269	GCF_001038685.1_concat_noambig_primer_00009
+Pba_ICMP_1526_concatenated_noambig	1848342	1848420	GCF_001038685.1_concat_noambig_primer_00010
