+++ tests/test_targets/pdp_primer3/subset/GCF_001038685.1_concat_noambig_named.eprimer3
@@ -1,0 +1,82 @@
+# EPRIMER3 PRIMERS tests/test_output/pdp_primer3/subset/GCF_001038685.1_concat_noambig_named.eprimer3 
+#                      Start  Len   Tm     GC%   Sequence
+# GCF_001038685.1_concat_noambig_primer_00001
+1    PRODUCT SIZE: 100
+     FORWARD PRIMER  914337     20   59.00  55.00  CGCGCTAGGGTCAAAGTTAC
+     REVERSE PRIMER  914416     20   58.99  50.00  TATTGTTGATCTGGCACGCG
+     INTERNAL OLIGO  0          0    0.00  0.00  
+
+
+
+# GCF_001038685.1_concat_noambig_primer_00002
+2    PRODUCT SIZE: 100
+     FORWARD PRIMER  2533415    20   58.99  50.00  CGATTCGCGGTTGAATTCCT
+     REVERSE PRIMER  2533494    20   59.01  55.00  CCTCGATGCTCACGTCAAAG
+     INTERNAL OLIGO  0          0    0.00  0.00  
+
+
+
+# GCF_001038685.1_concat_noambig_primer_00003
+3    PRODUCT SIZE: 100
+     FORWARD PRIMER  3896478    20   59.01  55.00  GGTCGTCGTGAGCTGAAATC
+     REVERSE PRIMER  3896557    20   58.99  50.00  GTGTGATAATCGGCGCTGTT
+     INTERNAL OLIGO  0          0    0.00  0.00  
+
+
+
+# GCF_001038685.1_concat_noambig_primer_00004
+4    PRODUCT SIZE: 100
+     FORWARD PRIMER  3467081    20   59.00  50.00  TGCTGACGAATCGCAATCAG
+     REVERSE PRIMER  3467160    20   58.99  50.00  ACGGTGAACACAAAGTGAGC
+     INTERNAL OLIGO  0          0    0.00  0.00  
+
+
+
+# GCF_001038685.1_concat_noambig_primer_00005
+5    PRODUCT SIZE: 100
+     FORWARD PRIMER  2807481    20   59.00  55.00  CAGTCGGGTAAGCATCGTTG
+     REVERSE PRIMER  2807560    20   59.02  55.00  GGTCTGGAACTGATGGGTCA
+     INTERNAL OLIGO  0          0    0.00  0.00  
+
+
+
+# GCF_001038685.1_concat_noambig_primer_00006
+6    PRODUCT SIZE: 100
+     FORWARD PRIMER  1210545    20   59.02  50.00  AAGGAGAAGACCGATGGCAA
+     REVERSE PRIMER  1210624    20   59.00  50.00  ATCCAGCCTGAAATCCACCA
+     INTERNAL OLIGO  0          0    0.00  0.00  
+
+
+
+# GCF_001038685.1_concat_noambig_primer_00007
+7    PRODUCT SIZE: 100
+     FORWARD PRIMER  445349     20   59.00  50.00  TGCCAGCCCAAGGAAGATAA
+     REVERSE PRIMER  445428     20   58.98  55.00  GGTCATGATGAAACTCGGCC
+     INTERNAL OLIGO  0          0    0.00  0.00  
+
+
+
+# GCF_001038685.1_concat_noambig_primer_00008
+8    PRODUCT SIZE: 100
+     FORWARD PRIMER  1585661    20   59.01  55.00  GCATGCTGTCAGTGACGTAG
+     REVERSE PRIMER  1585740    20   58.99  55.00  GGTATTGGGTAGCACGCTTG
+     INTERNAL OLIGO  0          0    0.00  0.00  
+
+
+
+# GCF_001038685.1_concat_noambig_primer_00009
+9    PRODUCT SIZE: 100
+     FORWARD PRIMER  3104191    20   58.98  55.00  GCGTGATTAGCCAAGAAGGG
+     REVERSE PRIMER  3104270    20   59.00  50.00  GTCGCATCTGCACGAGAAAT
+     INTERNAL OLIGO  0          0    0.00  0.00  
+
+
+
+# GCF_001038685.1_concat_noambig_primer_00010
+10   PRODUCT SIZE: 100
+     FORWARD PRIMER  1848342    20   59.01  50.00  GTGAAGCAAATCGCCAATGC
+     REVERSE PRIMER  1848421    20   58.99  55.00  CGGTTGTATAGGTTGCTGCC
+     INTERNAL OLIGO  0          0    0.00  0.00  
+
+
+
