+++ tests/test_targets/pdp_primer3/subset/GCF_001038685.1_concat_noambig_named.json
@@ -1,0 +1,1 @@
+[{"size": 100, "forward_seq": "CGCGCTAGGGTCAAAGTTAC", "forward_start": 914337, "forward_length": 20, "forward_tm": 59.003, "forward_gc": 55.0, "reverse_seq": "TATTGTTGATCTGGCACGCG", "reverse_start": 914416, "reverse_length": 20, "reverse_tm": 58.994, "reverse_gc": 50.0, "internal_seq": "", "internal_start": 0, "internal_length": 0, "internal_tm": 0.0, "internal_gc": 0.0, "name": "GCF_001038685.1_concat_noambig_primer_00001", "source": "tests/test_output/pdp_config/GCF_001038685.1_concat_noambig.fas", "sourcename": "Pba_ICMP_1526"}, {"size": 100, "forward_seq": "CGATTCGCGGTTGAATTCCT", "forward_start": 2533415, "forward_length": 20, "forward_tm": 58.993, "forward_gc": 50.0, "reverse_seq": "CCTCGATGCTCACGTCAAAG", "reverse_start": 2533494, "reverse_length": 20, "reverse_tm": 59.005, "reverse_gc": 55.0, "internal_seq": "", "internal_start": 0, "internal_length": 0, "internal_tm": 0.0, "internal_gc": 0.0, "name": "GCF_001038685.1_concat_noambig_primer_00002", "source": "tests/test_output/pdp_config/GCF_001038685.1_concat_noambig.fas", "sourcename": "Pba_ICMP_1526"}, {"size": 100, "forward_seq": "GGTCGTCGTGAGCTGAAATC", "forward_start": 3896478, "forward_length": 20, "forward_tm": 59.007, "forward_gc": 55.0, "reverse_seq": "GTGTGATAATCGGCGCTGTT", "reverse_start": 3896557, "reverse_length": 20, "reverse_tm": 58.993, "reverse_gc": 50.0, "internal_seq": "", "internal_start": 0, "internal_length": 0, "internal_tm": 0.0, "internal_gc": 0.0, "name": "GCF_001038685.1_concat_noambig_primer_00003", "source": "tests/test_output/pdp_config/GCF_001038685.1_concat_noambig.fas", "sourcename": "Pba_ICMP_1526"}, {"size": 100, "forward_seq": "TGCTGACGAATCGCAATCAG", "forward_start": 3467081, "forward_length": 20, "forward_tm": 58.998, "forward_gc": 50.0, "reverse_seq": "ACGGTGAACACAAAGTGAGC", "reverse_start": 3467160, "reverse_length": 20, "reverse_tm": 58.986, "reverse_gc": 50.0, "internal_seq": "", "internal_start": 0, "internal_length": 0, "internal_tm": 0.0, "internal_gc": 0.0, "name": "GCF_001038685.1_concat_noambig_primer_00004", "source": "tests/test_output/pdp_config/GCF_001038685.1_concat_noambig.fas", "sourcename": "Pba_ICMP_1526"}, {"size": 100, "forward_seq": "CAGTCGGGTAAGCATCGTTG", "forward_start": 2807481, "forward_length": 20, "forward_tm": 59.001, "forward_gc": 55.0, "reverse_seq": "GGTCTGGAACTGATGGGTCA", "reverse_start": 2807560, "reverse_length": 20, "reverse_tm": 59.016, "reverse_gc": 55.0, "internal_seq": "", "internal_start": 0, "internal_length": 0, "internal_tm": 0.0, "internal_gc": 0.0, "name": "GCF_001038685.1_concat_noambig_primer_00005", "source": "tests/test_output/pdp_config/GCF_001038685.1_concat_noambig.fas", "sourcename": "Pba_ICMP_1526"}, {"size": 100, "forward_seq": "AAGGAGAAGACCGATGGCAA", "forward_start": 1210545, "forward_length": 20, "forward_tm": 59.016, "forward_gc": 50.0, "reverse_seq": "ATCCAGCCTGAAATCCACCA", "reverse_start": 1210624, "reverse_length": 20, "reverse_tm": 58.998, "reverse_gc": 50.0, "internal_seq": "", "internal_start": 0, "internal_length": 0, "internal_tm": 0.0, "internal_gc": 0.0, "name": "GCF_001038685.1_concat_noambig_primer_00006", "source": "tests/test_output/pdp_config/GCF_001038685.1_concat_noambig.fas", "sourcename": "Pba_ICMP_1526"}, {"size": 100, "forward_seq": "TGCCAGCCCAAGGAAGATAA", "forward_start": 445349, "forward_length": 20, "forward_tm": 59.001, "forward_gc": 50.0, "reverse_seq": "GGTCATGATGAAACTCGGCC", "reverse_start": 445428, "reverse_length": 20, "reverse_tm": 58.983, "reverse_gc": 55.0, "internal_seq": "", "internal_start": 0, "internal_length": 0, "internal_tm": 0.0, "internal_gc": 0.0, "name": "GCF_001038685.1_concat_noambig_primer_00007", "source": "tests/test_output/pdp_config/GCF_001038685.1_concat_noambig.fas", "sourcename": "Pba_ICMP_1526"}, {"size": 100, "forward_seq": "GCATGCTGTCAGTGACGTAG", "forward_start": 1585661, "forward_length": 20, "forward_tm": 59.005, "forward_gc": 55.0, "reverse_seq": "GGTATTGGGTAGCACGCTTG", "reverse_start": 1585740, "reverse_length": 20, "reverse_tm": 58.986, "reverse_gc": 55.0, "internal_seq": "", "internal_start": 0, "internal_length": 0, "internal_tm": 0.0, "internal_gc": 0.0, "name": "GCF_001038685.1_concat_noambig_primer_00008", "source": "tests/test_output/pdp_config/GCF_001038685.1_concat_noambig.fas", "sourcename": "Pba_ICMP_1526"}, {"size": 100, "forward_seq": "GCGTGATTAGCCAAGAAGGG", "forward_start": 3104191, "forward_length": 20, "forward_tm": 58.983, "forward_gc": 55.0, "reverse_seq": "GTCGCATCTGCACGAGAAAT", "reverse_start": 3104270, "reverse_length": 20, "reverse_tm": 58.998, "reverse_gc": 50.0, "internal_seq": "", "internal_start": 0, "internal_length": 0, "internal_tm": 0.0, "internal_gc": 0.0, "name": "GCF_001038685.1_concat_noambig_primer_00009", "source": "tests/test_output/pdp_config/GCF_001038685.1_concat_noambig.fas", "sourcename": "Pba_ICMP_1526"}, {"size": 100, "forward_seq": "GTGAAGCAAATCGCCAATGC", "forward_start": 1848342, "forward_length": 20, "forward_tm": 59.006, "forward_gc": 50.0, "reverse_seq": "CGGTTGTATAGGTTGCTGCC", "reverse_start": 1848421, "reverse_length": 20, "reverse_tm": 58.986, "reverse_gc": 55.0, "internal_seq": "", "internal_start": 0, "internal_length": 0, "internal_tm": 0.0, "internal_gc": 0.0, "name": "GCF_001038685.1_concat_noambig_primer_00010", "source": "tests/test_output/pdp_config/GCF_001038685.1_concat_noambig.fas", "sourcename": "Pba_ICMP_1526"}]
