+++ tests/test_targets/pdp_primer3/subsetep3conf.json
@@ -1,0 +1,1 @@
+[{"features": "tests/test_output/pdp_filter/prodigal/GCF_000740965.1_concat.gff", "filestem": "GCF_000740965.1_concat", "filtered_seqfile": "tests/test_output/pdp_filter/prodigal/GCF_000740965.1_concat_prodigal.fas", "groups": ["Pectobacterium", "atrosepticum_NCBI", "gv1"], "name": "Pba_21A", "primers": "tests/test_output/pdp_primer3/subset/GCF_000740965.1_concat_named.json", "primersearch": null, "seqfile": "tests/test_output/pdp_config/GCF_000740965.1_concat.fas", "target_amplicons": null}, {"features": "tests/test_output/pdp_filter/prodigal/GCF_001038685.1_concat_noambig.gff", "filestem": "GCF_001038685.1_concat_noambig", "filtered_seqfile": "tests/test_output/pdp_filter/prodigal/GCF_001038685.1_concat_noambig_prodigal.fas", "groups": ["Pectobacterium", "atrosepticum_NCBI", "gv1"], "name": "Pba_ICMP_1526", "primers": "tests/test_output/pdp_primer3/subset/GCF_001038685.1_concat_noambig_named.json", "primersearch": null, "seqfile": "tests/test_output/pdp_config/GCF_001038685.1_concat_noambig.fas", "target_amplicons": null}, {"features": "tests/test_output/pdp_filter/prodigal/GCF_000011605.1.gff", "filestem": "GCF_000011605.1", "filtered_seqfile": "tests/test_output/pdp_filter/prodigal/GCF_000011605.1_prodigal.fasta", "groups": ["Pectobacterium", "atrosepticum_NCBI", "gv1"], "name": "Pba_SCRI1043", "primers": "tests/test_output/pdp_primer3/subset/GCF_000011605.1_named.json", "primersearch": null, "seqfile": "tests/test_input/sequences/GCF_000011605.1.fasta", "target_amplicons": null}]
